refector : 컴포넌트 분리 but 파일경로 문제
