+++ three session/test/src/card.jsx
@@ -1,4 +1,4 @@
-const card = [
+const cardImages = [
     { heat:1, src: "./img/가렌1.jpg", matched: false },
     { heat:2, src: "./img/다이애나1.jpg", matched: false },
     { heat:3, src: "./img/요네1.jpg", matched: false },
@@ -8,5 +8,5 @@ const card = [
     { heat:7, src: "./img/케이틀린.jpg", matched: false }
   ];
   
-  export default card;
+  export default cardImages;
